import-hmw/requirement.txt
google-cloud-storage==1.7.0
requests==2.13.0
configparser
pandas
numpy
google-oauth
google-cloud-bigquery
oauth2client
google-api-python-client
google-cloud
google-cloud-pubsub
